Move: Tests to a separate directory.
