+++ test/toctree.rst
@@ -1,0 +1,8 @@
+***********************
+``test/`` -- Unit tests
+***********************
+
+.. toctree::
+    :glob:
+
+    *.py
